POM 모델 적용
