+++ config.py
@@ -1,0 +1,5 @@
+APP_PATH = "C:\\Motion\\Motion_E\\Motion_E.exe"
+PROCESS_TITLE = "Motion_E.exe"
+MAX_RETRY = 3
+SCREENSHOT_SAVE_DIR = "FAIL"
+MOTION_VERSION_TITLE = "모션.ver"
