+++ locators/login_locators.py
@@ -1,0 +1,10 @@
+
+class LoginLocators:
+    LOGIN_FORM_TITLE = "로그인"
+    LOGIN_BUTTON_ID = "btnLogin"
+    ID_EDIT_ID = "txtUserID"
+    PW_EDIT_ID = "txtUserPW"
+    YKIHO_TEXT_ID = "txtYkiho"
+    YKIOH_CANCLE_BUTTON_ID = "btnCancle"
+    YKIHO_SAVE_BUTTON_ID = "btnSave"
+    SYSTEM_CANCLE_BUTTON_ID = "Close"
