+++ requirements.txt
@@ -1,0 +1,4 @@
+pywinauto 
+pyautogui 
+pillow 
+pyscreeze
